src/main/java/com/buyalskaya/day3/reader/DataReader.java
package com.buyalskaya.day3.reader;

import java.io.BufferedReader;
import java.io.FileReader;
import java.io.IOException;
import java.util.List;
import java.util.stream.Collectors;

public class DataReader {
    public List<String> readData(String filePath) {
        List<String> data;
        try (FileReader fileReader = new FileReader(filePath);
             BufferedReader bufferedReader = new BufferedReader(fileReader)) {
            data = bufferedReader.lines()
                    .collect(Collectors.toList());
        } catch (IOException e) {
            throw new RuntimeException("Error in opening file " + filePath, e);
        }
        return data;
    }
}
